QQuickEventPoint: always inform grabberItem on touch grab cancel
If we have a TouchCancel event we send it to the Item via sendEvent,
so it can be detected in the Item's event() or touchEvent() override.
If not, it still needs to be informed that it lost the grab, so
call touchUngrabEvent() instead.  MultiPointTouchArea for example
will stop showing that all its points are pressed, if the grab is
stolen by something else.  Followup to
59c753bc75c7cfd4068fbbba3c25e1f54c46f4c0

Change-Id: I211c02d51cdde8f8722600cf914485adcc2aa1f3
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>