tst_combobox: use larger window
test_down() needs to place the popup below the combobox button to be
able to test its pressed and down states while the popup is open. On
some high DPI displays (Chromebook Pixel) the popup did not fit fully
below, but got pushed partly over the button. Consequently, mouse
presses went to the popup instead of the button and the test failed.

Change-Id: I03e5768e00cf1eaf6d704a16ffe2b8d1705021b2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>