Slider: react immediately when using a mouse
The initial drag threshold is a necessary evil on touch to avoid
conflicting with flickables, but leads to bad experience (QTBUG-47081)
when using a mouse. Now that we have separate mouse and touch handling,
we can apply immediate moves when using a mouse, but keep the old
behavior on touch.

Changing the behavior is simple, but auto tests need quite many tweaks
so RangeSlider and Dial will be done separate follow up changes.

[ChangeLog][Important Behavior Changes] Sliders and Dials now react
immediately when using a mouse. Now the initial drag threshold applies
only on touch, to avoid conflicting with flickables.

Task-number: QTBUG-59920
Change-Id: Ifc1e6ed74c7894c3c854c975dff5238278827590
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>