Dial: react immediately when using a mouse
The initial drag threshold is a necessary evil on touch to avoid
conflicting with flickables, but leads to bad experience (QTBUG-47081)
when using a mouse. Now that we have separate mouse and touch handling,
we can apply immediate moves when using a mouse, but keep the old
behavior on touch.

Task-number: QTBUG-59920
Change-Id: I097cbdaed4933c0d907097b75f15ba16dcf615b1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>