RangeSlider: react immediately when using a mouse
The initial drag threshold is a necessary evil on touch to avoid
conflicting with flickables, but leads to bad experience (QTBUG-47081)
when using a mouse. Now that we have separate mouse and touch handling,
we can apply immediate moves when using a mouse, but keep the old
behavior on touch.

Task-number: QTBUG-59920
Change-Id: Ida36eed1f59a30173e3a2c53b55e044a87c98d8b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>