Speed up Qml library checksumming in developer builds
The use of sha1 to determine if there were any changes in the Qml
library in developer builds (for cache invalidation) works well, but it
results in timeouts when running tests on ARM as type compilation takes
too long. This is a developer feature and we might as well use a faster
hash such as Md5 that is sufficiently reliable for our purposes.

Change-Id: I917ae619c73a9cc821d888f16bfcccafeb8ebacc
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>