Micro-optimize QQuickIconLabel
Reduce indirection between the public and private objects. Things like
width, height, and componentComplete can be accessed directly in the
private object instead of asking them from the public object, which in
turn asks them from the private object. Furthermore, when direct access
is available on the stack, reference directly instead of repeatedly via
the d-ptr.

Change-Id: Idbb40bc86ebff7b7b22e21050941afd6f0027bfe
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>