ScrollBar: react immediately when using a mouse
The initial drag threshold is a necessary evil on touch to avoid
conflicting with flickables, but leads to bad experience (QTBUG-47081)
when using a mouse. Now that we have separate mouse and touch handling,
we can apply immediate moves when using a mouse, but keep the old
behavior on touch.

[ChangeLog][Important Behavior Changes] ScrollBar now reacts
immediately when using a mouse.

Task-number: QTBUG-58667
Change-Id: I51759fc01dc8b8536834a8a3cca635ff512f7dc9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>