TapHandler:wants: don't setPressed(false) unless pointId matches
This fixes the multibuttons manual test: the first button would see
that a second point was pressed, and that it didn't want it, because
it's outside the bounds.  This is not a good reason to discontinue
responding to the first touchpoint, which is the one it was first
interested in.

Change-Id: I7004a667873f235d3dda84b4261113d37d911763
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>