TapHandler: don't "want" every updated/stationary point
If the policy is ReleaseWithinBounds, we assumed wrongly that
wantsEventPoint() would not be asked about a point which
QQuickPointerSingleHandler::wantsEventPoint() would already rule out,
or about a point which was not grabbed either passively or exclusively.
But in fact it is asked about every point.  A tap is not going to
occur unless the press occurs within bounds; if that happened, then
QQuickPointerSingleHandler::m_pointId will have been set.  So a
TapHandler with this policy should remain interested in an updated
or stationary point only if it is the same one which it was already
interested in when it was pressed.

This gets the multipoint button example and autotest working.

Change-Id: I399c06071fd804cd6994d5f0153c307cec9d2f90
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>