QQPSingleHandler: verify wantsEventPoint() even when pointId is chosen
This allows TapHandler to veto the "wanting" during an update, if
the policy is violated.  For example, if the policy is WithinBounds,
and the point leaves the bounds, it will say that it does not want
the point; but without this change, it is not asked again when the
point is moved outside the bounds.

Change-Id: Id764b3d24966ce9c3c4795e9f2a0aa37821caa42
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>