QQPSingleHandler: don't accept a touchpoint other than the chosen one
This also caused a failure in the new TapHandler test:
wantsPointerEvent() can see stationary touchpoints, and they may end up
being "candidates" (although not if wantsEventPoint() returns false for
all stationary points), but don't get confused by that: we chose a
particular point, always the first for which wantsEventPoint() returns
true, not any subsequent points for which wantsEventPoint() returns true.
It's important to accept only that point, because in other places
where we don't have access to the return value from wantsEventPoint(),
we treat the accepted state as the _memory_ of what wantsEventPoint()
returned.  So it must be consistently the same.

Change-Id: Ia121bde50d956b45b2cf62eddf326293a5c02274
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>