QQuickPointerTouchEvent::reset(): preserve reordered touchpoints better
It's not just the grabbers we have to worry about.  If touchpoints in
one event are in a different order than in the previous one, we need
to also preserve anything that was remembered from when that touchpoint
was pressed: that is pressTimestamp, scenePressPos, and maybe sceneGrabPos.
So now we use a vector of structs for that purpose, which is more
extensible and easier to read.

Change-Id: Ibf2c0079693ed10988f0066184d53ee9106f2eca
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>