Fix performance regression in animators since 5.6
The costly addition is down to that we're performing TransformJob::postSync()
which we did not do before. This is superfluous though, as we do the same
bit of code in preSync() and there is no way the pointers we're checking
for can change between pre and post. Only missing bit is to move the invalidate().

Change-Id: I96420c14e11a5a33997ed6fac6a78695a917a8f2
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>