QQuickIcon: fix the type of the source property
For example, something like:

    Button {
        icon.source: "../images/drawer.png"
    }

did not work in the Gallery example, even though it works fine for
Image elements etc. The URL needs to be resolved relative to the
caller. By making the property an URL instead of a plain string,
the QML engine does the correct thing automatically.

Note: icon.source.length no longer works in auto tests, so it got
replaced by a compare().

Change-Id: I5fdb3aa55003d2f04e5458b7636e4dae3fa9857c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>