TestCase: Grow some new functionality on grabImage
Add some meta-information about the image that we can use to do slightly
stronger checks than just blindly poking pixel data (width & height
props).

In addition, when a test of a grabbed image fails (which they certainly do),
it is always nice if we can perform some diagnostics. To this end, add a
save() method to save the grabbed QImage to disk.

Tests for the new functionality are blacklisted on Linux at present, as they do
not yet appear stable, inheriting the other tst_grabImage flakiness.

[ChangeLog][QmlTest] The returned object from TestCase::grabImage now has
'width', 'height', and 'size' properties for additional verification. In
addition, there is a save() method to be able to persist the grabbed
image to disk (for diagnostics purposes, for example).

Change-Id: Ic651e0257102c514c39c3f648c05870f9d9c52e8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>