QSGBatchRenderer: Fix a performance regression noted by qmlbench
0a88774a0f11bf96a87012ac4a83e31ced19460b switched from using client side
pointers to VBOs, ostensibly for core profile support. Unfortunately,
some GPUs/drivers seem to suffer quite a bit with these, so only use
them if we must.

Results for changing_over_isolated_with_clip_rotated.qml:

eskil_linux_tx1:
    Before: Average: 65.6  frames; using  samples; MedianAll=65; StdDev=0.894427, CoV=0.0136346
    After:  Average: 120  frames; using  samples; MedianAll=120; StdDev=0, CoV=0

eskil_linux_focault (nvidia) has a similar regression in all branches
where the original commit is present, too:
    5.6 & v5.8.0 tag: 600
    5.8, 5.9 & dev branches: 399

Assuming this is the sole regression in that case and we end up at the
original 600 again, that would be a 50% gain there, and an 81% increase
on TX1.

Change-Id: I44af9e67698356200f9587e77c9409fdb756519d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>