QQuickDrawer: don't jump when dragging open
Typically drawer's drag margin is so small that one can barely notice
that it jumps when dragging begins. On systems with larger start drag
distance, or when drawer's drag margin is manually increased, the jump
was noticeable and disturbing.

Unfortunately several tests had to be adjusted accordingly, because
they were assuming that pressing at the edge and then dragging in the
middle would immediately snap the drawer. Now the tests are sending
a press at the edge, first moving past the start drag distance, and
then moving to the actual tested position.

Change-Id: Id8badf256fd4dd51f34db76ebe03bf6b15203cd9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>