DragHandler: allow parent to be different from target
The most obvious way to implement a Slider is to allow dragging the
knob - as on a real-world physical sliding potentiometer.  But to make
it easier on a touchscreen, it should be possible to touch anywhere
along the slider's travel, as on a QtQuick.Controls 2 Slider.  For
that purpose, we need to respond to events within the bounds of one
Item while actually dragging a different Item (the knob).  It's
similar to the way that PinchHandler can handle pinch gestures within
one Item while transforming another (which may be too small to get
both fingers inside).

Change-Id: Iac9a5f11a7a45e22d93fe52bf62d157c48d72d3d
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>