Simplify QQuickDrawerPrivate::startDrag()
The purpose is to calculate whether a press is within the drag margin
ie. whether a drawer can start dragging. There seems to be no reason to
call the clumsy QQuickWindowPrivate::dragOverThreshold() method, which
is different for mouse and touch events. In addition to the start drag
distance, it actually calculates start drag velocity too, which is
entirely irrelevant for us. Instead, we can do ourselves a very simple
mouse vs. touch agnostic calculation whether a press lands within the
drag margin.

Change-Id: I356a349fdef3df1455715ae7572dbcb9e8100f93
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>