QQuickRectangle: Use parenting on QQuickPen to speed up pen changes
We know that signal connections are slow, so let's not use them. QObject
has a parent pointer sitting there that looks nice and ripe, so:
let's set the parent pointer ourselves, and call update() via QQuickPen.

This improves delegates_rect_border.qml on my mbp by ~5%:
    Before: 394.6  frames; using  samples; MedianAll=395; StdDev=2.30217, CoV=0.00583419
    After:  417.4  frames; using  samples; MedianAll=417; StdDev=1.67332, CoV=0.00400891

... and should additionally decrease memory (no more signal connection).
Hopefully, the decrease in allocation and increase in performance might also
help with stability, since some systems seem a little flappy with this one.

Change-Id: I885654d606bc77d2949b9db81217426cf367b081
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>
Reviewed-by: Michael Brasser <michael.brasser@live.com>