QQuickRectangle: Optimize setGradient
Similarly to the QQuickPen optimization, we can avoid a signal
connection by using the parent pointer of the QQuickGradient.

This improves qmlbench's delegates_rect_gradient.qml by ~6% for me:
    Before: 300.6  frames; using  samples; MedianAll=301; StdDev=4.61519, CoV=0.01535
    After: 320.4  frames; using  samples; MedianAll=321; StdDev=1.94936, CoV=0.00608414

Again, hopefully, the resulting lessening in allocations and improvement
in performance will help stabilize the benchmark a little on some
systems, as well as the improvement to performance and memory use.

Change-Id: Id8e0399f5a4a0ef55d7fc9b8f100af229f389ddd
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>
Reviewed-by: Michael Brasser <michael.brasser@live.com>