QQuickMultiPointerHandler::wantsPointerEvent ignores grabs on press
A PinchHandler for example has strict requirements, so when they are
finally met, it needs to be willing to take over the grab from
something else.  This occurs when a point is pressed or released
(in which case a grabbing PointerHandler might probably give up its grab
anyway, if its requirements are not met anymore).  MPTA can typically
handle a variable number of points though, so it will not give
up its grabs, and PinchHandler must be able to steal from it.
The first step in stealing is for the handler to say that it wants
the points despite any existing grabs.

Change-Id: Ic3a2586d43e4cfe1877f90ce2e2862a9abc4b81b
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>