Fix memory leak in the memory manager
There was a bug in Chunk::sweep() that would lead
to parts of the memory not being freed as it should
have been.

This happened when an object 'overflowed' into the
next index of the estendBitmap, where we then would
not correctly clear the extends bits.

Fixes performance degradation in qmlbench's
delegates_item_empty_jscreation.qml over multiple runs.

Change-Id: Ia2cecd2ff218e4258a067a74631b5479589b7a7e
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>