Prevent PathView's parent stealing mouse grab during double-flick
This commit fixes an issue where mouse events could be stolen by the
parent of a PathView due to the PathView not correctly setting the
keep-mouse-grab flag in handleMousePressEvent().  This commit ensures
that the flag is correctly set, so that the second flick in a double
flick is handled by the PathView rather than being stolen by the
parent.

Task-number: QTBUG-59620
Change-Id: Iccdfe16e7e80e6d1d31f95c3dba9c8839b20f30f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>