Fix QQuickPopupPrivate::handleMouseEvent()
Touch moves are being abruptly stolen while flicking within a drawer.
Looks like there's a difference between mapping mouse events to scene
vs. using the window coordinates. All other places in the overlay and
popups are using window coordinates, so sync with that to ensure that
"drag over threshold" calculations are done based on press and move
coordinates that are in the same coordinate space.

Task-number: QTBUG-60521
Change-Id: Ied94870fe68a7e5da65b75d3daa585db6e5e9560
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>