QQuickIconImage: use QQuickPixmap::width/height instead of implicitSize
QQuickPixmap's implicitSize is the original size of the image on disk
before any scaling was done on it, which means that it doesn't account
for the scaling that QQuickImageProviderWithOptions::loadSize() does.
This was causing updateFillMode() to alternate between setting
PreserveAspectFit and Pad when the icon color was changed.

I was not able to come up with an auto test that reproduced the issue.

Task-number: QTBUG-60528
Change-Id: I1cc31f0c50875d794c37a6fc76ff0942e44d271a
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>