Optimizations to the generated byte code
Cut the size of the generated byte code in half, by storing
parameters in a more compact form and always storing the
instruction type in the instruction.

We can still used computed goto's for a fast interpreter, by
looking up the jump point for the next instruction in the jump
table.

Another advantage is that the byte code is now platform independent
(modulo endianness).

The change comes with a 3% performance impact on x86_64, which
is acceptable considering the size savings on the bytecode.

Change-Id: I37de3e1f94611987a85e65ea86536583aa965d6d
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>