TapHandler: don't give up passive grab on setPressed(false)
If a button with a TapHandler with the DragThreshold policy is inside
a Flickable, and you press on the button and start dragging, when you
exceed the drag threshold, TapHandler will not "want" the EventPoint
anymore.  That triggers it to call setPressed(false).  Flickable does
not have a grab because it was relying on filtering the children's
events; and QQuickWindow does not deliver touch updates to non-grabber
items.  So when TapHandler gives up its passive grab, Flickable does
not get a chance to filter TapHandler's touch update events anymore.
Thus, we cannot do that.  When the touchpoint is actually released
though, passive grabs are ungrabbed anyway.

This is required to get the new
tst_FlickableInterop::touchDragFlickableBehindButton()
autotest working.

Change-Id: I55a175ae358f75b9d7ab64f0b8124d91b6a9e1d6
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>