Enable build time QML caching for the Material & Universal styles
We had troubles enabling Qt Quick Compiler support for these styles,
because we wouldn't be able to select style-specific .qml files from
the resources of an unloaded/uninstantiated plugin. Build-time QML
caching doesn't seem to be a problem because the .qmlc files are not
built into resources but are available on the file system.

Task-number: QTBUG-58571
Change-Id: I93e7a82a72c4bcf3a92c7ca12186f24a22d4d342
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>