Software: Fix QQuickWidget rendering in QMdiArea
It is not safe to assume that the paintRegion is always the same as the
updateRegion reported by the software renderer, since in some cases more
needs to be flushed, such as in the case of the QMdiArea.  Now we make
sure to unite both of these regions and flush everything needed.

Task-number: QTBUG-60985
Change-Id: I0408c21e42dd4107b0974877144e8e93f2c30fae
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>